t/run-fio-tests: relax acceptance criteria for t0008
t0008 runs a 50/50 seq read/write job on a 32M file. Since a random number
generator determines whether fio issues a read or a write command, the
read/write ratio will not be exactly 50/50.  This patch relaxes the acceptance
criteria to avoid false positives when we do not use the default random number
seeds.

Signed-off-by: Vincent Fu <vincent.fu@samsung.com>