tools/fiograph: accommodate job files not ending in .fio
For job files not ending in .fio, fiograph will overwrite the job file
with a graphviz script and then delete it if --keep is not specified.

Fix this by creating temporary files to contain the graphviz script and
image file. Then rename or delete the script and image file as directed
by the user specified options.

Signed-off-by: Vincent Fu <vincent.fu@samsung.com>