Suppress sync engine QD > 1 warning if io_submit_mode is offload
The user is warned if iodepth > 1 when using a synchronous I/O engine,
since the engine can only submit one I/O at a time. This warning is not
accounting for the case of the user enabling I/O submission offload
threads via io_submit_mode=offload. Modified the warning conditional to
suppress the warning when this is the case.

Signed-off-by: Adam Horshack (horshack@live.com)