t/nvmept.py: test script for io_uring_cmd NVMe pass through
Test basic functionality of the io_uring_cmd ioengine with NVMe
pass-through commands.

This script starts with basic sequential and random read, write, trim,
readwrite, and trimwrite workloads. It also runs sequential and random
write workloads with verification. Finally it runs random read and write
workloads with almost all of the ioengine options enabled. The only
ioengine option not exercised is hipri because the NVMe driver must be
loaded with poll_queues set for this to work.

Since this is a destructive test the target device must be specified on
the command line.

Signed-off-by: Vincent Fu <vincent.fu@samsung.com>