client/server: enable per_job_logs option
On the client side log files were being overwritten when per_job_logs
was set to false because of the flags used when log files were opened.
Add per_job_logs to the on-wire protocol so that the client can adjust
the flags and open files in append mode when per_job_logs is set to
false.

Fixes: https://github.com/axboe/fio/issues/1032

Signed-off-by: Vincent Fu <vincent.fu@samsung.com>