fio: use thread flag count for ioengine flag shift
We store the thread and ioengine flags in the same unsigned long long
with the ioengine flags bit shifted by the number of thread flags. In
the macros related to the bit shift, just use the count of the number of
thread flags. This way any time we change the number of thread flags the
required shift will be automatically updated.

Signed-off-by: Vincent Fu <vincent.fu@samsung.com>